Move handler code to open source.
